Deprecate chunk buffer access in favour of BakedQuads for parts.
We used to rely on chunk buffer access due to Vanilla lighting not being good enough for microblocks. However, this is no longer the case, vanilla produces nearly identical, if not the same, lighting in all cases.

With the popularity of Mods like Sodium, these chunk buffer access patches CCL needs to maintain are becoming more problematic to maintain.